Swap out syslog libraries for better output formatting
Imports srslog as a syslog library to allow for better and
more standard syslog output formatting.

https://github.com/RackSec/srslog

The go standard syslog library produces a syslog message that is
a mix of rfc3164 and 5424 formats. We'd like an rfc3164 compliant
message without the hostname, since we're only sending messages
locally (to syslog).